+++ CHANGELOG.md
@@ -22,6 +22,22 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+- Crash trying to decode non https URLs from the database (jpg5)
+
+### Security
+
 ## [9.10.3] - 2026-05-19
 
 ### Fixed